기존 안쓰는 객체 삭제 및 클래스 이름 리팩토링
