+++ src/main/java/nextstep/ladder/domain/PlayLadderGame.java
@@ -8,68 +8,22 @@ public class PlayLadderGame {
     public Users generateResultsForAllPlayers(GameInfo paramGameInfo, Ladder paramLadder) {
         Users users = paramGameInfo.getUsers();
         List<User> userList = users.getUsers();
-        List<Line> ladder = paramLadder.getLadder();
+        List<LadderLine> ladder = paramLadder.getLadder();
         List<Result> results = paramGameInfo.getResults().getResults();
 
         for (int i = 0; i < users.getCountOfPerson(); i++) {
-            int index = initIndex(users.getCountOfPerson(), i);
-            index = repeatLineForFindIndex(ladder, users.getCountOfPerson() - 1, index);
+            int index = i;
+            index = getIndex(ladder, index);
             userList.get(i).setResult(results.get(index).getResult());
         }
         return new Users(userList);
     }
 
-    private int plusIndex(Point point) {
-        return point.plusIndex();
-    }
-
-    private int minusIndex(Point point) {
-        return point.minusIndex();
-    }
-
-    private int judgeIndex(Point point, boolean isLeftLine) {
-        return point.judgeIndex(isLeftLine);
-    }
-
-    private int initIndex(int userCount, int i) {
-        int index = i;
-        if (i >= userCount) {
-            index = userCount - 1;
-        }
-        return index;
-    }
-
-    private int repeatLineForFindIndex(List<Line> ladder, int last, int index) {
+    private int getIndex(List<LadderLine> ladder, int index) {
         for (int j = 0; j < ladder.size(); j++) {
-            List<Point> points = ladder.get(j).getPoints();
-            Point point = getPoint(points, last, index);
-            point.setIndex(index);
-            index = getIndex(last, index, points, point);
+            index = ladder.get(j).move(index);
         }
         return index;
-    }
-
-    private int getIndex(int last, int index, List<Point> points, Point point) {
-        if (index == last) {
-            index = minusIndex(point);
-            return index;
-        }
-        if (index == 0) {
-            index = plusIndex(point);
-            return index;
-        }
-        if (index > 0 && index < last) {
-            Point prePoint = points.get(index - 1);
-            index = judgeIndex(point, prePoint.isNextPoint());
-        }
-        return index;
-    }
-
-    private Point getPoint(List<Point> points, int last, int index) {
-        if (index < last) {
-            return points.get(index);
-        }
-        return points.get(index - 1);
     }
 
     public static PlayLadderGame getPlayLadderGame() {
