+++ src/test/java/nextstep/ladder/view/ResultViewTest.java
@@ -2,7 +2,6 @@ package nextstep.ladder.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -11,17 +10,13 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import com.sun.tools.internal.xjc.reader.xmlschema.bindinfo.BIConversion.Static;
-
+import nextstep.ladder.domain.Direction;
 import nextstep.ladder.domain.GameInfo;
+import nextstep.ladder.domain.ImprovingPoint;
 import nextstep.ladder.domain.Ladder;
-import nextstep.ladder.domain.Line;
+import nextstep.ladder.domain.LadderLine;
 import nextstep.ladder.domain.PlayLadderGame;
-import nextstep.ladder.domain.Point;
-import nextstep.ladder.domain.Result;
-import nextstep.ladder.domain.Results;
 import nextstep.ladder.domain.ResultsTest;
-import nextstep.ladder.domain.User;
 import nextstep.ladder.domain.Users;
 import nextstep.ladder.domain.UsersTest;
 
@@ -29,8 +24,9 @@ class ResultViewTest {
     private static ResultView resultView = ResultView.getResultView();
     private static PlayLadderGame playLadderGame = PlayLadderGame.getPlayLadderGame();
 
-    private Line line;
-    private List<Line> lineList;
+    private LadderLine line;
+    private LadderLine line2;
+    private List<LadderLine> lineList;
     private Ladder ladder;
 
     private Users resultUsers;
@@ -38,13 +34,21 @@ class ResultViewTest {
 
     @BeforeEach
     void setUp() {
-        List<Point> points = Stream.of(new Point(true),
-                                       new Point(false),
-                                       new Point(true))
-                                   .collect(Collectors.toList());
+        List<ImprovingPoint> points = Stream.of(new ImprovingPoint(0, Direction.of(false, true)),
+                                                new ImprovingPoint(1, Direction.of(true, false)),
+                                                new ImprovingPoint(2, Direction.of(false, true)),
+                                                new ImprovingPoint(3, Direction.of(true, false)))
+                                            .collect(Collectors.toList());
 
-        line = new Line(points);
-        lineList = Stream.of(line, line)
+        List<ImprovingPoint> points2 = Stream.of(new ImprovingPoint(0, Direction.of(false, false)),
+                                                 new ImprovingPoint(1, Direction.of(false, true)),
+                                                 new ImprovingPoint(2, Direction.of(true, false)),
+                                                 new ImprovingPoint(3, Direction.of(false, false)))
+                                             .collect(Collectors.toList());
+
+        line = new LadderLine(points);
+        line2 = new LadderLine(points2);
+        lineList = Stream.of(line, line2)
                          .collect(Collectors.toList());
 
         ladder = new Ladder(lineList);
@@ -63,7 +67,7 @@ class ResultViewTest {
     @Test
     void appendLine() {
         String line = resultView.appendLine(this.line);
-        assertThat(line).isEqualTo("     |-----|     |-----|");
+        assertThat(line).isEqualTo("     |-----|     |-----|     ");
     }
 
     @DisplayName("실행결과를 출력한다.")
@@ -78,8 +82,8 @@ class ResultViewTest {
     void drawLadder() {
         String drawLadder = resultView.drawLadder(gameInfo, ladder);
         assertThat(drawLadder).isEqualTo(" pobi honux crong    jk\n"
-                                         + "     |-----|     |-----|\n"
-                                         + "     |-----|     |-----|\n"
+                                         + "     |-----|     |-----|     \n"
+                                         + "     |     |-----|     |     \n"
                                          + "    꽝  5000     꽝  3000");
 
     }
@@ -88,15 +92,16 @@ class ResultViewTest {
     @Test
     void playResult() {
         String resultByUser = resultView.playResultByUser(resultUsers, "honux");
-        assertThat(resultByUser).isEqualTo("5000");
+        assertThat(resultByUser).isEqualTo("꽝");
     }
 
+    @DisplayName("전체 실행결과를 출력한다.")
     @Test
     void playAllResult() {
         String playAllResult = resultView.playAllResult(resultUsers);
         assertThat(playAllResult).isEqualTo("pobi : 꽝\n"
-                                            + "honux : 5000\n"
-                                            + "crong : 꽝\n"
-                                            + "jk : 3000");
+                                            + "honux : 꽝\n"
+                                            + "crong : 3000\n"
+                                            + "jk : 5000");
     }
 }
